requirements.txt
pyotp
smartapi-python
websocket-client
omspy
oauthlib
webdriver_manager
